AI_CornerDetect/requirements.txt
opencv-python
pyyaml
numpy
torch
torchvision
tqdm
matplotlib
